C++: weird algorithm

+++ elementary_computer_science/C_Cpp/weird_algorithm.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-	int n;
+	long long n;
 	cin >> n;
 	while (1) {
 		cout << n << " ";
